feat(CLI): Set key counter
Adds new CLI command (and RPC support with backend implementation) for
manually setting the key counters for generation and symmetric
addresses.

Under normal circumstances, the wallet handles the counters well.
However it is possible that during an irregular sync or wallet import,
the key counters become out of sync with the transactions that the
wallet received. In this case, the wallet is aware of UTXOs that
it cannot spend, because none of the spending keys it has derived so
far are capable of unlocking the UTXO. In this case, manually
overriding the key counter can make the UTXOs spendable again.